fix backup table actions
